test-detection/requirements.txt
boto3
gitpython